# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "gem",
        "description": "gem is project about transportation",
        "version": "1"
    },
    "paths": {
        "/api/addresses": {
            "get": {
                "tags": [
                    "addresses"
                ],
                "summary": "get all addresses",
                "description": "get all addresses",
                "operationId": "dc6b190d3f07db7ed032fa181dc45150",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "addresses"
                ],
                "summary": "create address",
                "description": "create address",
                "operationId": "6344fc023028e43a5490f1c6372bd391",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sender_reciver_id": {
                                        "description": "foreign key to senders_recivers table",
                                        "type": "integer"
                                    },
                                    "province_code": {
                                        "type": "integer"
                                    },
                                    "province_title": {
                                        "type": "string"
                                    },
                                    "city_code": {
                                        "type": "integer"
                                    },
                                    "city_title": {
                                        "type": "string"
                                    },
                                    "postal_code": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/addresses/{address}": {
            "get": {
                "tags": [
                    "addresses"
                ],
                "summary": "get address by id",
                "description": "get address by id",
                "operationId": "85431c549062d0c537257448ad3eedda",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "addresses"
                ],
                "summary": "delete address ",
                "description": "delete address ",
                "operationId": "fc2684b35fd67400b287e6b5638d75b2",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "addresses"
                ],
                "summary": "update address",
                "description": "update address",
                "operationId": "8c7a1bda21bd34b28826fb6c6a9bcf12",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sender_reciver_id": {
                                        "description": "foreign key to senders_recivers table",
                                        "type": "integer"
                                    },
                                    "province_code": {
                                        "type": "integer"
                                    },
                                    "province_title": {
                                        "type": "string"
                                    },
                                    "city_code": {
                                        "type": "integer"
                                    },
                                    "city_title": {
                                        "type": "string"
                                    },
                                    "postal_code": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/addresses/load-address-by-postal-code": {
            "post": {
                "tags": [
                    "addresses"
                ],
                "summary": "loading address by postal code",
                "description": "loading address by postal code",
                "operationId": "4f601287d15776a4674b7ae1acee05b5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "postal_code": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    }
}